Removed unnecessary variables, constructors, pauses.
