+++ src/main/java/core/basesyntax/Excavator.java
@@ -1,0 +1,18 @@
+package core.basesyntax;
+
+public class Excavator extends Machine {
+
+    public Excavator(String name) {
+        super(name);
+    }
+
+    @Override
+    public void doWork() {
+        System.out.println(getName() + " Start");
+    }
+
+    @Override
+    public void stopWork() {
+        System.out.println(getName() + " Stop");
+    }
+}
