+++ src/main/java/core/basesyntax/Machine.java
@@ -1,0 +1,21 @@
+package core.basesyntax;
+
+abstract class Machine {
+    private String name;
+
+    public Machine(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public abstract void doWork();
+
+    public abstract void stopWork();
+}
